requirements.txt
torch
torchvision
torchaudio
numpy
scikit-learn
h5py
matplotlib
tqdm
pyyaml
open3d
